develop

+++ packages/server/database/migrations/1735401994854_create_social_class_economical_situation_per_turns_table.ts
@@ -7,7 +7,7 @@ export default class extends BaseSchema {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id');
 
-      table.integer('amount').notNullable();
+      table.float('amount').notNullable();
       table.integer('turn').notNullable();
 
       table
